refactor + make it change the rendering in inventory
