--- src/main/java/com/github/noamm9/mixin/MixinItemInHand.java
@@ -1,27 +1,0 @@
-package com.github.noamm9.mixin;
-
-import com.github.noamm9.features.impl.visual.RevertAxes;
-import net.minecraft.client.renderer.ItemInHandRenderer;
-import net.minecraft.world.item.ItemStack;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
-
-import java.util.Objects;
-
-@Mixin(ItemInHandRenderer.class)
-public class MixinItemInHand {
-    @ModifyVariable(
-            method = "renderArmWithItem",
-            at = @At("HEAD"),
-            argsOnly = true
-    )
-    private ItemStack revertAxe(ItemStack original) {
-        if (original == null || original.isEmpty()) {
-            return original;
-        }
-
-        ItemStack replacement = RevertAxes.INSTANCE.shouldReplace(original);
-        return Objects.requireNonNullElse(replacement, original);
-    }
-}
