+++ src/main/resources/noammaddons.mixins.json
@@ -53,11 +53,10 @@
     "MixinGameRenderer",
     "MixinGui",
     "MixinGuiGraphics",
-    "MixinItemInHand",
     "MixinItemInHandRenderer",
+    "MixinItemModelResolver",
     "MixinLerpingBossEvent",
     "MixinLightTexture",
-    "MixinLivingEntityItem",
     "MixinMinecraft",
     "MixinMouseHandler",
     "MixinMultiPlayerGameMode",
